main: workaround running inside a "/" less chroot (like mock)
This commit makes image-builder-cli transparently deal with running
inside a chroot as mock creates them. The issue here is that
bwrap does not like it when it runs without a "/".  This is
based on the excellent work from jlebon in blubblewrap issue

This commit is probably overcomplicated for what we need, it
depends on how much control we have in the koji plugin. This
commit assume we have zero control and can just run "image-builder"
so it does its setup and re-execs itself. This works fine *but*
is complicated.

The alternative is to just provide a wrapper script in the
package that is run when we run in koji.