main: add `--bootc-ref` as a thin wrapper around calling bib
This commit adds the ability to create bootc content with
`image-builder`. The usage is like an ostree image, i.e. it
is run via:
```console
```

Internally this currently just runs `bib` from the official bib
container and translates the options. However this is just a first
step but the UI for the user should not change even if all the
code from bib moves into ibcli.