progress: error if there are not enough priviledges to run osbuild
This commit detects if there are enough priviledges to run osbuild
and if not errors. This prepares for the `supermin` build mode
which will be used in the followup commits when not enough
priviledges are detected.

It currently checks for uid==0 and CAP_SYS_ADMIN, this is (probably)
a reasonable approximation. We could also check for more like seccomp
filtering or selinux "unconfined" label.