main: add "depsolve" output
This commit adds a new `depsolve` subcommand that can be
used to get a full depsolve result for the given image
type and build config. Note that this is inherentily
racy unless it is used together with `manifest` and
the resulting manifest is used (see below). This is
because on the next manifest generation the repositories
may change yielding different results.

There is going to be a `--with-depsolve` flag for
manifest to also write the depsolve result as part of
manifest creation (just like --with-sbom). This will
avoid the above race.